// src/app/models/zonacomun.ts
export class zonaComun {

    id: number = 0;
    idCuenta: number = 0;
    nombre: string = "";
    capacidad: number = 0;
    idCategoria: number = 0;
    foto:string="";
    estado: number = 0;
    descripcion: string = "";
    noPermitido: string = "";
    disponibilidad: string = "";
    date: string = "";
    constructor() {
    }
}